project/requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.1
python-dotenv>=1.0
gunicorn>=22.0
Pillow>=11.0
django-admin-sortable2>=2.2
django-ckeditor>=6.7.1
